+++ backend/app/Filament/Resources/NotificationTemplateResource/Pages/EditNotificationTemplate.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\NotificationTemplateResource\Pages;
+
+use App\Filament\Resources\NotificationTemplateResource;
+use Filament\Resources\Pages\EditRecord;
+
+class EditNotificationTemplate extends EditRecord
+{
+    protected static string $resource = NotificationTemplateResource::class;
+}